Revert "refactor apis (#53)"
This reverts commit 3962d9094afb08ff354900c1c671bbb23bd546a9.
