--- research_app/__about__.py
@@ -1,22 +1,0 @@
-__version__ = "0.0.1"
-__author__ = "PyTorchLightning et al."
-__author_email__ = "name@pytorchlightning.ai"
-__license__ = "TBD"
-__copyright__ = f"Copyright (c) 2021-2022, {__author__}."
-__homepage__ = "https://github.com/PyTorchLightning/lightning-template-research-app"
-__docs__ = "Project x"
-__long_doc__ = """
-What is it?
------------
-
-"""
-
-__all__ = [
-    "__author__",
-    "__author_email__",
-    "__copyright__",
-    "__docs__",
-    "__homepage__",
-    "__license__",
-    "__version__",
-]
