+++ research_app/components/poster_work.py
