+++ setup.py
@@ -1,51 +1,19 @@
-#!/usr/bin/env python
+import setuptools
 
-import os
-from importlib.util import module_from_spec, spec_from_file_location
-
-from setuptools import find_packages, setup
-
-_PATH_ROOT = os.path.dirname(__file__)
-
-
-def _load_py_module(fname, pkg="research_app"):
-    spec = spec_from_file_location(os.path.join(pkg, fname), os.path.join(_PATH_ROOT, pkg, fname))
-    py = module_from_spec(spec)
-    spec.loader.exec_module(py)
-    return py
-
-
-about = _load_py_module("__about__.py")
-setup_tools = _load_py_module("setup_tools.py")
-long_description = setup_tools._load_readme_description(_PATH_ROOT, homepage=about.__homepage__, ver=about.__version__)
-
-# https://packaging.python.org/discussions/install-requires-vs-requirements /
-# keep the meta-data here for simplicity in reading this file... it's not obvious
-# what happens and to non-engineers they won't know to look in init ...
-# the goal of the project is simplicity for researchers, don't want to add too much
-# engineer specific practices
-
-setup(
+setuptools.setup(
     name="research_app",
-    version=about.__version__,
-    description=about.__docs__,
-    author=about.__author__,
-    author_email=about.__author_email__,
-    url=about.__homepage__,
-    download_url="https://github.com/PyTorchLightning/lightning",
-    license=about.__license__,
-    packages=find_packages(exclude=["tests", "docs"]),
-    long_description=long_description,
-    long_description_content_type="text/markdown",
-    include_package_data=True,
-    zip_safe=False,
-    keywords=["deep learning", "pytorch", "AI"],
-    python_requires=">=3.6",
-    entry_points={
-        "console_scripts": [
-            "lightning = lightning.cli.lightning_cli:main",
-        ],
+    version="0.0.1",
+    description="Research App",
+    author="Grid.ai",
+    packages=setuptools.find_packages(where="research_app*"),
+    include_package_data=False,
+    extras_require={
+        "dev": [
+            "ipdb==0.13.9",
+            "black==22.3.0",
+            "isort==5.10.1",
+            "pre-commit==2.18.1",
+        ],  # E501
     },
-    setup_requires=["wheel"],
-    install_requires=setup_tools._load_requirements(_PATH_ROOT),
+    python_requires=">=3.8",
 )
